@@ -4,13 +4,15 @@ import { AuthService } from '../../Core/Service/Implements/AuthService';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 
+
 @Component({
   selector: 'app-menu-nav-bar',
   standalone: true,
   imports: [RouterOutlet,
     FormsModule,
     CommonModule,
-    RouterLink],
+    RouterLink,
+    ],
   templateUrl: './menu-nav-bar.component.html',
   styleUrl: './menu-nav-bar.component.css',
   
@@ -19,6 +21,7 @@ export class MenuNavBarComponent {
 
   title = 'anjadefr';
   isHomePage: boolean = false;
+  isNavbarCollapsed = true;
 
   constructor(private router: Router,private authService: AuthService) {}
   ngOnInit() {
@@ -35,6 +38,9 @@ export class MenuNavBarComponent {
     }
     const validRoles = ['presidente', 'comisionados', 'secretarias','administrador'];
     return validRoles.includes(user.usuariorol.descripcion) && user.estadoCuenta.estado === 'activo';
+  }
+  collapseNavbar() {
+    this.isNavbarCollapsed = true;
   }
   downloadEstatutos(){
     // Ruta del documento Word
